# Optimise `to_ascii_lowercase_smolstr`, `to_ascii_uppercase_smolstr`