Optimisation opportunity highlighted by testing against `String` perf and using benchmarks in #96.

## PR bench
```
group                                  master                  faster-ascii-case
-----                                  ------                  -----------------
to_ascii_lowercase_smolstr len=12      2.06     13.5±0.01ns    1.00      6.6±0.01ns
to_ascii_lowercase_smolstr len=50      4.31    110.3±0.19ns    1.00     25.6±0.07ns
to_ascii_lowercase_smolstr len=1000    22.17  1396.8±2.63ns    1.00     63.0±0.01ns
```

The change provides a significant perf improvement on inline & heap reprs. Also faster than `String` for len=12 now :slightly_smiling_face: